+++ packages/akashic-cli-install/src/cli.ts
@@ -1,6 +1,6 @@
 import * as fs from "fs";
 import * as path from "path";
-import * as commander from "commander";
+import { Command } from "commander";
 import { ConsoleLogger, CliConfigurationFile, CliConfigInstall } from "@akashic/akashic-cli-commons";
 import { promiseInstall } from "./install";
 
@@ -19,6 +19,7 @@ function cli(param: CliConfigInstall): void {
 
 var ver = JSON.parse(fs.readFileSync(path.resolve(__dirname, "..", "package.json"), "utf8")).version;
 
+const commander = new Command();
 commander
 	.version(ver);
 
@@ -33,7 +34,8 @@ commander
 
 export function run(argv: string[]): void {
 	commander.parse(argv);
-	CliConfigurationFile.read(path.join(commander["cwd"] || process.cwd(), "akashic.config.js"), (error, configuration) => {
+	const options = commander.opts();
+	CliConfigurationFile.read(path.join(options["cwd"] || process.cwd(), "akashic.config.js"), (error, configuration) => {
 		if (error) {
 			console.error(error);
 			process.exit(1);
@@ -42,11 +44,11 @@ export function run(argv: string[]): void {
 		const conf = configuration.commandOptions.install || {};
 		cli({
 			args: commander["args"] ?? conf.args,
-			cwd: commander["cwd"] ?? conf.cwd,
-			link: commander["link"] ?? conf.link,
-			quiet: commander["quiet"] ?? conf.quiet,
-			plugin: commander["plugin"] ?? conf.plugin,
-			omitPackagejson: commander["omitPackagejson"] ?? conf.omitPackagejson
+			cwd: options["cwd"] ?? conf.cwd,
+			link: options["link"] ?? conf.link,
+			quiet: options["quiet"] ?? conf.quiet,
+			plugin: options["plugin"] ?? conf.plugin,
+			omitPackagejson: options["omitPackagejson"] ?? conf.omitPackagejson
 		});
 	});
 }