+++ packages/akashic-cli-modify/src/cli.ts
@@ -1,8 +1,10 @@
 import * as fs from "fs";
 import * as path from "path";
-import * as commander from "commander";
+import { Command } from "commander";
 import { ConsoleLogger, CliConfigurationFile, CliConfigModify } from "@akashic/akashic-cli-commons";
 import { promiseModifyBasicParameter } from "./modify";
+
+const commander = new Command();
 
 function cliBasicParameter(target: string, value: string, opts: CliConfigModify): void {
 	var logger = new ConsoleLogger({ quiet: opts.quiet });
@@ -21,7 +23,7 @@ function defineCommand(commandName: string): void {
 		.option("-C, --cwd <dir>", "The directory incluedes game.json")
 		.option("-q, --quiet", "Suppress output")
 		.action((value: string, opts: CliConfigModify = {}) => {
-			CliConfigurationFile.read(path.join(commander["cwd"] || process.cwd(), "akashic.config.js"), (error, configuration) => {
+			CliConfigurationFile.read(path.join(opts["cwd"] || process.cwd(), "akashic.config.js"), (error, configuration) => {
 				if (error) {
 					console.error(error);
 					process.exit(1);