+++ packages/akashic-cli-serve/src/client/atsumaru/ScoreboardData.ts
@@ -1,4 +1,4 @@
-// RPGアツマールのスコアボードの情報
+// ゲームアツマールのスコアボードの情報
 export interface ScoreboardData {
 	myRecord: null | {
 	  isNewRecord: boolean,